refactor(coral): remove msw dependency in dataSetup
