+++ coral/src/app/features/topics/acl-request/TopicAclRequest_producer.test.tsx
@@ -5,16 +5,11 @@ import { userEvent } from "@testing-library/user-event";
 import { Route, Routes } from "react-router-dom";
 import TopicAclRequest from "src/app/features/topics/acl-request/TopicAclRequest";
 import { mockCreateAclRequest } from "src/domain/acl/acl-api-msw";
-import { getMockedResponseGetClusterInfoFromEnvironment } from "src/domain/cluster/cluster-api.msw";
-import {
-  mockGetClusterInfoFromEnv,
-  mockgetAllEnvironmentsForTopicAndAcl,
-} from "src/domain/environment/environment-api.msw";
+import { getClusterInfoFromEnvironment } from "src/domain/cluster";
+import { getMockedResponseGetClusterInfoFromEnvironment } from "src/domain/cluster/cluster-api-test-helper";
+import { getAllEnvironmentsForTopicAndAcl } from "src/domain/environment";
 import { createMockEnvironmentDTO } from "src/domain/environment/environment-test-helper";
-import {
-  mockGetTopicNames,
-  mockGetTopicTeam,
-} from "src/domain/topic/topic-api.msw";
+import { getTopicNames, getTopicTeam } from "src/domain/topic";
 import {
   mockedResponseTopicNames,
   mockedResponseTopicTeamLiteral,
@@ -40,39 +35,46 @@ jest.mock("@aivenio/aquarium", () => ({
   useToast: () => mockedUseToast,
 }));
 
+jest.mock("src/domain/environment/environment-api.ts");
+const mockgetAllEnvironmentsForTopicAndAcl =
+  getAllEnvironmentsForTopicAndAcl as jest.MockedFunction<
+    typeof getAllEnvironmentsForTopicAndAcl
+  >;
+
+jest.mock("src/domain/topic/topic-api.ts");
+const mockGetTopicNames = getTopicNames as jest.MockedFunction<
+  typeof getTopicNames
+>;
+const mockGetTopicTeam = getTopicTeam as jest.MockedFunction<
+  typeof getTopicTeam
+>;
+
+jest.mock("src/domain/cluster/cluster-api.ts");
+const mockGetClusterInfoFromEnv =
+  getClusterInfoFromEnvironment as jest.MockedFunction<
+    typeof getClusterInfoFromEnvironment
+  >;
+
 const dataSetup = ({ isAivenCluster }: { isAivenCluster: boolean }) => {
-  mockgetAllEnvironmentsForTopicAndAcl({
-    mswInstance: server,
-    response: {
-      data: [
-        createMockEnvironmentDTO({
-          name: "TST",
-          id: "1",
-        }),
-        createMockEnvironmentDTO({
-          name: "DEV",
-          id: "2",
-        }),
-        createMockEnvironmentDTO({
-          name: "PROD",
-          id: "3",
-        }),
-      ],
-    },
-  });
-  mockGetTopicNames({
-    mswInstance: server,
-    response: mockedResponseTopicNames,
-  });
-  mockGetTopicTeam({
-    mswInstance: server,
-    response: mockedResponseTopicTeamLiteral,
-    topicName: "aivtopic1",
-  });
-  mockGetClusterInfoFromEnv({
-    mswInstance: server,
-    response: getMockedResponseGetClusterInfoFromEnvironment(isAivenCluster),
-  });
+  mockgetAllEnvironmentsForTopicAndAcl.mockResolvedValue([
+    createMockEnvironmentDTO({
+      name: "TST",
+      id: "1",
+    }),
+    createMockEnvironmentDTO({
+      name: "DEV",
+      id: "2",
+    }),
+    createMockEnvironmentDTO({
+      name: "PROD",
+      id: "3",
+    }),
+  ]);
+  mockGetTopicNames.mockResolvedValue(mockedResponseTopicNames);
+  mockGetTopicTeam.mockResolvedValue(mockedResponseTopicTeamLiteral);
+  mockGetClusterInfoFromEnv.mockResolvedValue(
+    getMockedResponseGetClusterInfoFromEnvironment(isAivenCluster)
+  );
 };
 
 const assertSkeleton = async () => {
